Added first revision of code with upload
Arduino side of the actuator control. Developing using only one actuator for now, so code is being done on an Arduino Uno. So far, this code contains logic to turn on the IBT-2 motor controller and detect if the limit switch buttons are being pressed so movement in their direction will stop.